WorkflowRun: Fix job.queued_duration_second
It's not really that useful to have the time from when a job starts
util the time when the first step starts. That's always 0 or close to
0.

Instead, let's track the time of when the workflow starts until the
first step of the job is started. This better represents how long the
job has been queued.